# ai-version/doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todo List API",
    "description": "A simple CRUD API for managing a todo list, backed by an in-memory array.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Task",
      "description": "Operations related to todo tasks"
    }
  ],
  "paths": {
    "/task": {
      "get": {
        "tags": ["Task"],
        "summary": "Get the list of tasks",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Task"],
        "summary": "Create a new task",
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Title is missing or empty",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/task/{id}": {
      "get": {
        "tags": ["Task"],
        "summary": "Get a specific task by id",
        "operationId": "getTaskById",
        "parameters": [
          { "$ref": "#/components/parameters/TaskId" }
        ],
        "responses": {
          "200": {
            "description": "The requested task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Task"],
        "summary": "Replace a task's title and done status",
        "operationId": "updateTask",
        "parameters": [
          { "$ref": "#/components/parameters/TaskId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskUpdateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Body is invalid or empty",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Task"],
        "summary": "Delete a task",
        "operationId": "deleteTask",
        "parameters": [
          { "$ref": "#/components/parameters/TaskId" }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "TaskId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Numeric id of the task",
        "schema": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Buy groceries"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        },
        "required": ["id", "title", "done"]
      },
      "TaskCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Buy groceries"
          }
        },
        "required": ["title"]
      },
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Buy groceries and milk"
          },
          "done": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["title", "done"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Task with id 5 not found"
          }
        }
      }
    }
  }
}